Moved reload_module functions to the core plugin file to ensure latest are loaded
